Move multistream related classes to the corresponding package
