+++ src/main/kotlin/io/libp2p/core/mux/StreamMuxer.kt
@@ -2,8 +2,8 @@ package io.libp2p.core.mux
 
 import io.libp2p.core.Stream
 import io.libp2p.core.StreamHandler
-import io.libp2p.core.protocol.ProtocolBinding
-import io.libp2p.core.protocol.ProtocolBindingInitializer
+import io.libp2p.core.multistream.ProtocolBinding
+import io.libp2p.core.multistream.ProtocolBindingInitializer
 import java.util.concurrent.CompletableFuture
 
 interface StreamMuxer : ProtocolBinding<StreamMuxer.Session> {
